Use `swift_add_tool()` for unit test suite (#374)
* Remove default-declared and defaultable copy constructors

These aren't needed.

* Mark unused arguments with new ALBATROSS_UNUSED

This can be replaced by `[[maybe_unused]]` if/when C++17 becomes a requirement.

* Be more careful about casting and integral types

 - Introduce safe casting functions which assert if precision is lost

 - Use safe casting functions to do conversions

 - Use `Eigen::Index` for indexing Eigen types and `std::size_t` for
   indexing STL types

 - A few other small changes to e.g. function argument types where
    things didn't line up

* Fix various instances of name shadowing

* Suppress errors from -Wshadow due to GCC 6 bug

These don't appear on any other compiler or version.

* Use `swift_add_tool()` for test suite

+++ tests/test_samplers.cc
@@ -138,7 +138,7 @@ TEST(test_samplers, test_samplers_gp) {
 }
 
 inline long int get_group(const double &f) {
-  return static_cast<double>(floor(f / 5.));
+  return static_cast<long int>(floor(f / 5.));
 }
 
 struct LeaveOneIntervalOut {
